fixup! Reuse simpletxidstore
