+++ platform/common/utils/iterators.go
@@ -1,0 +1,8 @@
+package utils
+
+type Iterator[V any] interface {
+	Next() (V, error)
+	Close()
+}
+
+func Map[A any, B any](iterator Iterator[A], transformer func(A) (B, error)) Iterator[B] {}
