--- platform/fabric/core/generic/vault/txidstore/ftxid.proto
@@ -1,17 +1,0 @@
-syntax = "proto3";
-
-option go_package = "github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/vault/txidstore";
-
-package txidstore;
-
-message ByNum {
-    string txid = 1;
-    int32 code = 2;
-    string message = 3;
-}
-
-message ByTxid {
-    uint64 pos = 1;
-    int32 code = 2;
-    string message = 3;
-}
